Split stream
