+++ frictionless/plugins/stream/__init__.py
@@ -1,0 +1,3 @@
+from .plugin import StreamPlugin
+from .control import StreamControl
+from .loader import StreamLoader
